backend/requirement.txt
django==4.0
django-filter
django-cors-headers
djangorestframework
djangorestframework_simplejwt
markdown
